Create a text_recognition.py file and move three adadpters from text_detection.py there, meanwhile fixing the too many lines error
