+++ tools/accuracy_checker/accuracy_checker/adapters/text_detection.py
@@ -668,147 +668,6 @@ class TextProposalConnector:
         return text_recs
 
 
-class LPRAdapter(Adapter):
-    __provider__ = 'lpr'
-    prediction_types = (CharacterRecognitionPrediction,)
-
-    def process(self, raw, identifiers=None, frame_meta=None):
-        if not self.label_map:
-            raise ConfigError('LPR adapter requires dataset label map for correct decoding.')
-        raw_output = self._extract_predictions(raw, frame_meta)
-        predictions = raw_output[self.output_blob]
-        result = []
-        for identifier, output in zip(identifiers, predictions):
-            decoded_out = self.decode(output.reshape(-1))
-            result.append(CharacterRecognitionPrediction(identifier, decoded_out))
-
-        return result
-
-    def decode(self, outputs):
-        decode_out = str()
-        for output in outputs:
-            if output == -1:
-                break
-            decode_out += str(self.label_map[int(output)])
-
-        return decode_out
-
-
-class BeamSearchDecoder(Adapter):
-    __provider__ = 'beam_search_decoder'
-    prediction_types = (CharacterRecognitionPrediction, )
-
-    @classmethod
-    def parameters(cls):
-        parameters = super().parameters()
-        parameters.update({
-            'beam_size': NumberField(
-                optional=True, value_type=int, min_value=1, default=10,
-                description="Size of the beam to use during decoding."
-            ),
-            'blank_label': NumberField(
-                optional=True, value_type=int, min_value=0, description="Index of the CTC blank label."
-            ),
-            'softmaxed_probabilities': BoolField(
-                optional=True, default=False, description="Indicator that model uses softmax for output layer "
-            )
-        })
-        return parameters
-
-    def validate_config(self):
-        super().validate_config(on_extra_argument=ConfigValidator.IGNORE_ON_EXTRA_ARGUMENT)
-        self.beam_size = self.get_value_from_config('beam_size')
-        self.blank_label = self.launcher_config.get('blank_label')
-        self.softmaxed_probabilities = self.get_value_from_config('softmaxed_probabilities')
-
-    def process(self, raw, identifiers=None, frame_meta=None):
-        if not self.label_map:
-            raise ConfigError('Beam Search Decoder requires dataset label map for correct decoding.')
-        if self.blank_label is None:
-            self.blank_label = len(self.label_map)
-        raw_output = self._extract_predictions(raw, frame_meta)
-        output = raw_output[self.output_blob]
-        output = np.swapaxes(output, 0, 1)
-
-        result = []
-        for identifier, data in zip(identifiers, output):
-            if self.softmaxed_probabilities:
-                data = np.log(data)
-            seq = self.decode(data, self.beam_size, self.blank_label)
-            decoded = ''.join(str(self.label_map[char]) for char in seq)
-            result.append(CharacterRecognitionPrediction(identifier, decoded))
-        return result
-
-    @staticmethod
-    def decode(probabilities, beam_size=10, blank_id=None):
-        """
-         Decode given output probabilities to sequence of labels.
-        Arguments:
-            probabilities: The output log probabilities for each time step.
-            Should be an array of shape (time x output dim).
-            beam_size (int): Size of the beam to use during decoding.
-            blank_id (int): Index of the CTC blank label.
-        Returns the output label sequence.
-        """
-        def make_new_beam():
-            return defaultdict(lambda: (-np.inf, -np.inf))
-
-        def log_sum_exp(*args):
-            if all(a == -np.inf for a in args):
-                return -np.inf
-            a_max = np.max(args)
-            lsp = np.log(np.sum(np.exp(a - a_max) for a in args))
-
-            return a_max + lsp
-
-        times, symbols = probabilities.shape
-        # Initialize the beam with the empty sequence, a probability of 1 for ending in blank
-        # and zero for ending in non-blank (in log space).
-        beam = [(tuple(), (0.0, -np.inf))]
-
-        for time in range(times):
-            # A default dictionary to store the next step candidates.
-            next_beam = make_new_beam()
-
-            for symbol_id in range(symbols):
-                current_prob = probabilities[time, symbol_id]
-
-                for prefix, (prob_blank, prob_non_blank) in beam:
-                    # If propose a blank the prefix doesn't change.
-                    # Only the probability of ending in blank gets updated.
-                    if symbol_id == blank_id:
-                        next_prob_blank, next_prob_non_blank = next_beam[prefix]
-                        next_prob_blank = log_sum_exp(
-                            next_prob_blank, prob_blank + current_prob, prob_non_blank + current_prob
-                        )
-                        next_beam[prefix] = (next_prob_blank, next_prob_non_blank)
-                        continue
-                    # Extend the prefix by the new character symbol and add it to the beam.
-                    # Only the probability of not ending in blank gets updated.
-                    end_t = prefix[-1] if prefix else None
-                    next_prefix = prefix + (symbol_id,)
-                    next_prob_blank, next_prob_non_blank = next_beam[next_prefix]
-                    if symbol_id != end_t:
-                        next_prob_non_blank = log_sum_exp(
-                            next_prob_non_blank, prob_blank + current_prob, prob_non_blank + current_prob
-                        )
-                    else:
-                        # Don't include the previous probability of not ending in blank (prob_non_blank) if symbol
-                        #  is repeated at the end. The CTC algorithm merges characters not separated by a blank.
-                        next_prob_non_blank = log_sum_exp(next_prob_non_blank, prob_blank + current_prob)
-
-                    next_beam[next_prefix] = (next_prob_blank, next_prob_non_blank)
-                    # If symbol is repeated at the end also update the unchanged prefix. This is the merging case.
-                    if symbol_id == end_t:
-                        next_prob_blank, next_prob_non_blank = next_beam[prefix]
-                        next_prob_non_blank = log_sum_exp(next_prob_non_blank, prob_non_blank + current_prob)
-                        next_beam[prefix] = (next_prob_blank, next_prob_non_blank)
-
-            beam = sorted(next_beam.items(), key=lambda x: log_sum_exp(*x[1]), reverse=True)[:beam_size]
-        best = beam[0]
-        return best[0]
-
-
 class EASTTextDetectionAdapter(Adapter):
     __provider__ = 'east_text_detection'
 
@@ -998,72 +857,3 @@ class EASTTextDetectionAdapter(Adapter):
             new_p_1 = np.zeros((0, 4, 2))
 
         return np.concatenate([new_p_0, new_p_1])
-
-
-class CTCGreedySearchDecoder(Adapter):
-    __provider__ = 'ctc_greedy_search_decoder'
-    prediction_types = (CharacterRecognitionPrediction, )
-
-    @classmethod
-    def parameters(cls):
-        parameters = super().parameters()
-        parameters.update({
-            'blank_label': NumberField(
-                optional=True, value_type=int, min_value=0, default=0, description="Index of the CTC blank label."
-            )
-        })
-        return parameters
-
-    def validate_config(self):
-        super().validate_config(on_extra_argument=ConfigValidator.IGNORE_ON_EXTRA_ARGUMENT)
-        self.blank_label = self.launcher_config.get('blank_label')
-
-    def process(self, raw, identifiers=None, frame_meta=None):
-        if not self.label_map:
-            raise ConfigError('CTCGreedy Search Decoder requires dataset label map for correct decoding.')
-        if self.blank_label is None:
-            self.blank_label = 0
-        raw_output = self._extract_predictions(raw, frame_meta)
-        output = raw_output[self.output_blob]
-        preds_index = np.argmax(output, 2)
-        preds_index = preds_index.transpose(1, 0)
-
-        result = []
-        for identifier, data in zip(identifiers, preds_index):
-            data = data[None, :]
-            seq = self.decode(data, self.blank_label)
-            decoded = ''.join(str(self.label_map[char]) for char in seq)
-            result.append(CharacterRecognitionPrediction(identifier, decoded))
-
-        return result
-
-    @staticmethod
-    def decode(probabilities, blank_id=None):
-        """
-         Decode given output probabilities to sequence of labels.
-        Arguments:
-            probabilities: The output log probabilities for each time step.
-            blank_id (int): Index of the CTC blank label.
-        Returns the output label sequence.
-        """
-        preds_sizes = np.array([probabilities.shape[1]] * probabilities.shape[0])
-        preds_index = probabilities.reshape(-1)
-
-        text_index = preds_index
-        length = preds_sizes
-
-        char_list = []
-        index = 0
-        for l in length:
-            t = text_index[index:index + l]
-
-            # NOTE: t might be zero size
-            if t.shape[0] == 0:
-                continue
-
-            for i in range(l):
-                # removing repeated characters and blank.
-                if t[i] != blank_id and (not (i > blank_id and t[i - 1] == t[i])):
-                    char_list.append(t[i])
-            index += l
-        return char_list
